Avoid generating duplicate files using the templates

+++ lib/watirmark-gli/new.rb
@@ -16,15 +16,27 @@ module WatirmarkGli
     end
 
     def self.create_directory(directory_path)
-      FileUtils.mkdir_p directory_path
+      if Dir.exists? directory_path
+        puts "Directory already exist at path:'#{directory_path}'"
+      else
+        FileUtils.mkdir_p directory_path
+      end
     end
 
     def self.erb_config_yml
       file_name_erb = "#{@templates_path}/config.yml.erb"
       file_name = "#{@relative_path}/config.yml"
-      template_file = File.open(file_name_erb, 'r').read
-      erb = ERB.new(template_file)
-      File.open(file_name, 'w+') { |file| file.write(erb.result(binding)) }
+      process_erb_file(file_name_erb, file_name)
+    end
+
+    def self.process_erb_file(file_name_erb, file_name)
+      if File.exist?(file_name)
+        puts "File already exist at path:'#{file_name}'"
+      else
+        template_file = File.open(file_name_erb, 'r').read
+        erb = ERB.new(template_file)
+        File.open(file_name, 'w+') { |file| file.write(erb.result(binding)) }
+      end
     end
   end
 end
